requirements.txt
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
idna==2.10
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
python-dotenv==0.14.0
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
validators==0.18.0
wcwidth==0.2.5
